run_executor_worker/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==3.4.7
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.14.0
grpcio==1.62.2
grpcio-health-checking==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.6
nodeenv==1.8.0
./openai-1.26.0-py3-none-any.whl
pika==1.3.2
platformdirs==4.2.2
pre-commit==3.7.1
protobuf==4.25.3
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
validators==0.28.0
virtualenv==20.26.2
watchfiles==0.21.0
weaviate-client==4.5.6
